Shell: Added 'clear' command to reset the shell to empty

+++ src/kernel/shell.c
@@ -76,6 +76,11 @@ int shell_run(arena* _kernel_arena, ui_box_t* box) {
 
             switch (r.code) {
             case ERROR_NONE: {
+                if (r.is_clear_command) {
+                    ui_scroll_vertical_n(box, current_line);
+                    current_line = 0;
+                    break;
+                }
                 for (int i = 0; i < r.string_result_count; i++) {
                     output_linecount += ui_putstr(box, 2, current_line + output_linecount, r.string_results[i]);
                 }
@@ -146,6 +151,7 @@ struct builtin_result shell_process(char* buf) {
     if (strcmp(ctx.args[0], "newf") == 0)   return shell_newf_builtin(&ctx);
     if (strcmp(ctx.args[0], "delf") == 0)   return shell_delf_builtin(&ctx);
     if (strcmp(ctx.args[0], "appf") == 0)   return shell_appf_builtin(&ctx);
+    if (strcmp(ctx.args[0], "clear") == 0)  return (struct builtin_result) {.is_clear_command = true};
     if (strcmp(ctx.args[0], "reboot") == 0) sys_reboot();
 
     return BUILTIN_RESULT_SUC_NO_MSG(ERROR_INVALID_CMD); // invalid command supplied
@@ -161,11 +167,13 @@ struct builtin_result shell_read_builtin(const struct argument_ctx* arg_ctx) {
     // 1. Convert input filename into a FAT12 compatible 8.3 format
     char name_8_3[12] = "           ";
     if (fat_drive_filename_to_8_3(arg_ctx->args[1], name_8_3) == -1) {
+        log_group_end("ShellReadBuiltin");
         return BUILTIN_RESULT_SUC_NO_MSG(ERROR_GENERIC);
     }
 
     // 2. Search the file system for the 8.3 name
     if (!(f = fat_drive_find_file(name_8_3))) {
+        log_group_end("ShellReadBuiltin");
         return BUILTIN_RESULT_SUC_NO_MSG(ERROR_FILE_NOT_FOUND);
     }
 
@@ -173,6 +181,7 @@ struct builtin_result shell_read_builtin(const struct argument_ctx* arg_ctx) {
     buf = arena_alloc(kernel_arena, f->Size);
     if (!fat_drive_read_file(f, buf)) {
         log_info("Status", "Failed to read file");
+        log_group_end("ShellReadBuiltin");
         return BUILTIN_RESULT_SUC(ERROR_FILE_READ, 0);
     }
     log_group_end("ShellReadBuiltin");
@@ -207,7 +216,7 @@ struct builtin_result shell_newf_builtin(const struct argument_ctx* arg_ctx) {
     log_info("8.3", "\"%s\"", name_8_3);
     
     log_group_end("Shell 'newf'");
-    return BUILTIN_RESULT_ERR(ERROR_UNIMPLEMENTED, "Not implemented");
+    return BUILTIN_RESULT_ERR(ERROR_UNIMPLEMENTED, "WIP");
 }
 
 struct builtin_result shell_appf_builtin(const struct argument_ctx* arg_ctx) {
